requirements.txt
asgiref==3.8.1
Django==5.2
gunicorn==23.0.0
numpy==2.2.5
packaging==25.0
pillow==11.2.1
sqlparse==0.5.3
whitenoise==6.9.0
